requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
discord.py==1.1.1
idna==2.8
more-itertools==7.0.0
multidict==4.5.2
pep8==1.7.1
pluggy==0.11.0
py==1.8.0
pytest==4.5.0
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
wcwidth==0.1.7
websockets==6.0
yarl==1.3.0
pyshorteners==0.6.1

